+++ R/metrics.R
@@ -6,8 +6,9 @@
 #' @param object A 'nestcv.glmnet', 'nestcv.train', 'nestcv.SuperLearner' or
 #'   'outercv' object.
 #' @param extra Logical whether additional performance metrics are gathered for
-#'   binary classification models: area under precision recall curve (PR.AUC),
-#'   Cohen's kappa, F1 score, Matthew's correlation coefficient (MCC).
+#'   classification models: area under precision recall curve (PR.AUC, binary
+#'   classification only), Cohen's kappa, F1 score, Matthews correlation
+#'   coefficient (MCC).
 #' @param innerCV Whether to calculate metrics for inner CV folds. Only
 #'   available for 'nestcv.glmnet' and 'nestcv.train' objects.
 #' @param positive For binary classification, either an integer 1 or 2 for the
@@ -17,7 +18,16 @@
 #' @details
 #' Area under precision recall curve is estimated by trapezoidal estimation 
 #' using `MLmetrics::PRAUC()`.
+#' 
+#' For multi-class classification models, Matthews correlation coefficient is
+#' calculated using Gorodkin's method. Multi-class F1 score (macro F1) is
+#' calculated as the arithmetic mean of the class-wise F1 scores.
 #' @returns A named numeric vector of performance metrics.
+#' @references
+#' Gorodkin, J. (2004). \emph{Comparing two K-category assignments by a
+#' K-category correlation coefficient}. Computational Biology and Chemistry. 28
+#' (5): 367–374.
+#' @seealso [mcc()]
 #' @export
 #'
 metrics <- function(object, extra = FALSE, innerCV = FALSE, positive = 2) {
@@ -39,16 +49,29 @@ metrics <- function(object, extra = FALSE, innerCV = FALSE, positive = 2) {
     return(object$summary)
   }
   met <- object$summary$metrics
-  if (extra && nlevels(object$y) == 2) {
-    # binary classification
-    aucpr <- prc(object, positive = positive)$auc
-    tab <- object$summary$table
-    mcc <- mcc(tab)
-    if (is.numeric(positive)) positive <- colnames(tab)[positive]
-    ccm <- caret::confusionMatrix(tab, mode = "everything", positive = positive)
-    extra <- setNames(c(aucpr, ccm$overall["Kappa"], ccm$byClass["F1"], mcc), 
-                      c("PR.AUC", "Kappa", "F1", "MCC"))
-    met <- c(met, extra)
+  if (extra) {
+    if (nlevels(object$y) == 2) {
+      # binary classification
+      aucpr <- prc(object, positive = positive)$auc
+      tab <- object$summary$table
+      mcc <- mcc(tab)
+      if (is.numeric(positive)) positive <- colnames(tab)[positive]
+      ccm <- caret::confusionMatrix(tab, mode = "everything", positive = positive)
+      extra <- setNames(c(aucpr, ccm$overall["Kappa"], ccm$byClass["F1"], mcc), 
+                        c("PR.AUC", "Kappa", "F1", "MCC"))
+      met <- c(met, extra)
+    } else if (nlevels(object$y) > 2) {
+      # multiclass
+      tab <- object$summary$table
+      mcc <- mcc_multi(tab)
+      if (is.numeric(positive)) positive <- colnames(tab)[positive]
+      ccm <- caret::confusionMatrix(tab, mode = "everything", positive = positive)
+      f1 <- ccm$byClass[, "F1"]
+      f1.macro <- mean(f1)
+      extra <- setNames(c(ccm$overall["Kappa"], f1.macro, mcc),
+                        c("Kappa", "F1.macro", "MCC"))
+      met <- c(met, extra)
+    }
   }
   if (innerCV && inherits(object, c("nestcv.glmnet", "nestcv.train"))) {
     inner_met <- innercv_summary(object)$metrics
@@ -86,7 +109,25 @@ nvar <- function(object) {
 }
 
 
-# Matthew's correlation coefficient
+#' Matthews correlation coefficient
+#' 
+#' Calculates Matthews correlation coefficient (MCC) which is in essence a
+#' correlation coefficient between the observed and predicted binary
+#' classifications. It has also been generalised to multi-class classification.
+#' 
+#' @param cm A contingency table or matrix of predicted vs observed classes with
+#'   reference classes in columns and predicted classes in rows.
+#' @details Use `mcc()` for 2x2 tables (binary classification). `mcc_multi()` is
+#'   for multi-class classification with k x k tables and is calculated using
+#'   Gorodkin's method.
+#' @returns Returns a value between -1 and +1. A coefficient of +1
+#' represents a perfect prediction, 0 no better than random prediction and -1
+#' indicates total disagreement between prediction and observation.
+#' @references
+#' Gorodkin, J. (2004). \emph{Comparing two K-category assignments by a
+#' K-category correlation coefficient}. Computational Biology and Chemistry. 28
+#' (5): 367–374.
+#' @export
 mcc <- function(cm) {
   tp <- cm[2, 2]
   tn <- cm[1, 1]
@@ -94,4 +135,14 @@ mcc <- function(cm) {
   fn <- cm[1, 2]
   (tp*tn - fp*fn) /
     sqrt((tp+fp) * (tp+fn) * (tn+fp) * (tn+fn))
+}
+
+#' @rdname mcc
+#' @export
+mcc_multi <- function(cm) {
+  N <- sum(cm)
+  tt <- colSums(cm)
+  pp <- rowSums(cm)
+  RK <- (N * sum(diag(cm)) - sum(tt * pp)) / (sqrt(N^2 - sum(tt^2)) * sqrt(N^2 - sum(pp^2)))
+  RK
 }